+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed deprecated EXOS methods
+- Removed deprecated snapshot shipping methods. Use backup shipping instead
 
 ## [1.25.3] - 2025-04-10